feat:exclude icons folder from type definition
This commit excludes the icons folder from type definition generation in .

The reason for this change is that the icons folder contains SVG files, which are not TypeScript code and therefore should not be included in the type definition file. By excluding the icons folder, we can reduce the size of the type definition file and improve the performance of TypeScript compilation.