feat: Implement icons using LazyLoad
This commit implements the use of lazy loading for icons, which improves performance by deferring the loading of icon components until they are needed.

Previously, all icons were loaded eagerly, which could lead to slower page load times, especially on pages with a large number of icons.

With this change, icons are now only loaded when they are actually used on a page. This is accomplished by using the React Suspense and lazy APIs.

Changelog:

- Feat: Implement icons using LazyLoad